Separate functionality in smaller methods to build readable code
