+++ subclasses/create_book.rb
@@ -1,15 +1,20 @@
-require_relative '../book'
+require_relative '../classes/book.rb'
 
 class CreateBook
-    
+
+    def title_selection
+      print 'Title: '
+      title = gets.chomp
+    end
+
+    def author_selection
+      print 'Author: '
+      author = gets.chomp
+    end
+
     def create_book(books)
-        print 'Title: '
-        title = gets.chomp
-        print 'Author: '
-        author = gets.chomp
-    
-        book = Book.new(title, author)
+        book = Book.new(title_selection, author_selection)
         books << book unless books.include?(book)
         puts 'Book created successfully'
-      end
+    end
 end
