+++ subclasses/create_rentals.rb
@@ -1,21 +1,28 @@
-require_relative '../rental'
+require_relative '../classes/rental.rb'
 
 class CreateRental
-    
-def create_rental(books, people, rentals)
+
+  def book_selection(books)
     puts 'Select a book from the following list by number'
     books.each_with_index { |book, index| puts "#{index}) Title: \"#{book.title}\", Author: #{book.author}" }
-    number_book = gets.chomp
-    puts
-    puts 'Select a person from the following list by number(not id)'
+    number_book = gets.chomp.to_i
+  end
+
+  def person_selection(people)
+    puts "\nSelect a person from the following list by number(not id)"
     people.each_with_index do |person, index|
       puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
     end
-    number_people = gets.chomp
-    puts
-    print 'Date: '
+    number_people = gets.chomp.to_i
+  end
+
+  def date_selection
+    print "\nDate: "
     date = gets.chomp
-    rental = Rental.new(date, people[number_people.to_i], books[number_book.to_i])
+  end
+    
+  def create_rental(books, people, rentals)
+    rental = Rental.new(date_selection, people[person_selection(people)], books[book_selection(books)])
     rentals << rental unless rentals.include?(rental)   
   end
 end
